Issue/3207 add attributes part1

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/ProductGlobalAttribute.kt
@@ -14,7 +14,7 @@ data class ProductGlobalAttribute(
     val orderBy: String,
     val hasArchives: Boolean,
     val termsId: String,
-    val remoteId: Int
+    val remoteId: Long
 ) : Parcelable {
     fun toDataModel(cachedAttribute: WCGlobalAttributeModel? = null): WCGlobalAttributeModel {
         return (cachedAttribute ?: WCGlobalAttributeModel()).also {
@@ -26,7 +26,7 @@ data class ProductGlobalAttribute(
             it.orderBy = orderBy
             it.hasArchives = hasArchives
             it.termsId = termsId
-            it.remoteId = remoteId
+            it.remoteId = remoteId.toInt()
         }
     }
 }
@@ -34,7 +34,7 @@ data class ProductGlobalAttribute(
 fun WCGlobalAttributeModel.toAppModel(): ProductGlobalAttribute {
     return ProductGlobalAttribute(
         id = this.id,
-        remoteId = this.remoteId,
+        remoteId = this.remoteId.toLong(),
         localSiteId = this.localSiteId,
         name = this.name,
         slug = this.slug,
